refactor: move blocked applications management to a dedicated ClipboardPage preference section
